requirements.txt
# Google ADKとAI関連
google-cloud-aiplatform>=1.38.0
google-genai>=1.0.0

# 環境管理
python-dotenv>=1.0.0

# 非同期処理
aiohttp>=3.9.0
asyncio>=3.4.3

# データ処理
pydantic>=2.5.0

# ロギング
structlog>=23.2.0
